- Header에 sticky 처리 App.tsx의 header로 이동 - ZoneDetailHeader의 Info를 ZoneDetailPage에서 하도록 위치 이동 - ZoneSearchResultHeaderContainer, HeaderPage.scss 제거
