--- src/pages/HeaderPage/HeaderPage.scss
@@ -1,3 +1,0 @@
-.header {
-  background-color: var(--BackgroundColor);
-}
