+++ src/pages/HeaderPage/HeaderPage.tsx
@@ -1,38 +1,21 @@
-import React from 'react';
-import './HeaderPage.scss';
-import { withRouter, RouteComponentProps } from 'react-router-dom';
-
-import queryString from 'query-string';
+import React from "react";
+import { useLocation } from "react-router-dom";
 
 //Containers
-import DefaultHeaderContainer from '../../containers/HeaderContainer/DefaultHeaderContainer';
-import ZoneDetailHeaderContainer from '../../containers/HeaderContainer/ZoneDetailHeaderContainer';
-import ZoneSearchResultHeaderContainer from '../../containers/HeaderContainer/ZoneSearchResultHeaderContainer';
+import DefaultHeaderContainer from "../../containers/HeaderContainer/DefaultHeaderContainer";
+import ZoneDetailHeaderContainer from "../../containers/HeaderContainer/ZoneDetailHeaderContainer";
 
-type ParamsType = {
-  id: string;
-  feature: string;
-}
-
-const HeaderPage = ({ match, location, history}: RouteComponentProps<ParamsType>) => {  
-  
-  const query = queryString.parse(location.search);
-
+const HeaderPage = () => {
+  const location = useLocation();
   let HeaderContainer;
 
-  if (location.hash.includes('/zone/')) {
+  if (location.hash.includes("/zone/")) {
     HeaderContainer = ZoneDetailHeaderContainer;
-  } else if (location.search) {
-    HeaderContainer = ZoneSearchResultHeaderContainer;
   } else {
     HeaderContainer = DefaultHeaderContainer;
   }
 
-  return (
-    <div className="header">
-      <HeaderContainer />
-    </div>
-  )
-}
+  return <HeaderContainer />;
+};
 
-export default withRouter(HeaderPage);
+export default HeaderPage;
